pyproject.toml
[project]
name = "gemini-agent-mcp"
version = "0.1.0"
description = "AI Agent System with Streamlit, FastAPI, MCP, and Google Gemini"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.30.0",
    "openai>=1.10.0",
    "mcp>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "starlette",  # 👈 여기에 콤마(,)가 꼭 있어야 합니다!

    # 새로 추가한 부분
    "chromadb>=0.4.22",             # 👈 여기도 콤마
    "sentence-transformers>=2.3.1"  # 마지막은 없어도 되지만 붙여도 됨
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]